+++ src/Addons/34thPRC_ArmourCustom/data/odst2/config_weapons.hpp
@@ -1,3 +1,4 @@
 #include "yandere\config_weapons.hpp"
 #include "cappy\config_weapons.hpp"
+#include "ambrose\config_weapons.hpp"
 #include "shewolf\config_weapons.hpp"