+++ src/changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Ambrose Custom Armour
 - Grenadier for Marines/ODSTs
 - Scout for marines /ODSTs
 - Shewolf Custom Helmet